tetragon: Move read_exe into process.h header
Moving read_exe into process.h header so it can be easily used in
following changes from other places.
